requirements.txt
# PPT审查工具依赖包

# 核心依赖
python-pptx>=0.6.21
PyYAML>=6.0
requests>=2.28.0

# 可选依赖（用于打包）
pyinstaller>=5.0.0

# 开发依赖
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
